+++ tests/modules/test_blazeMeterUploader.py
@@ -1,4 +1,5 @@
 import logging
+import math
 import os
 import shutil
 from io import BytesIO
@@ -6,8 +7,9 @@ import time
 
 from bzt.modules.aggregator import DataPoint, KPISet
 from tests import BZTestCase, random_datapoint, __dir__
-from bzt.six import URLError
+from bzt.six import URLError, iteritems, viewvalues
 from bzt.modules.blazemeter import BlazeMeterUploader, BlazeMeterClient, BlazeMeterClientEmul, ResultsFromBZA
+from bzt.modules.blazemeter import MonitoringBuffer
 from tests.mocks import EngineEmul
 import bzt.modules.blazemeter
 
@@ -69,6 +71,19 @@ class TestBlazeMeterUploader(BZTestCase):
         obj = BlazeMeterClient(logging.getLogger(''))
         obj.address = "https://a.blazemeter.com"
         obj.ping()
+
+    def test_monitoring_buffer_limit_option(self):
+        obj = BlazeMeterUploader()
+        obj.engine = EngineEmul()
+        obj.client = BlazeMeterClientEmul(logging.getLogger(''))
+        obj.client.results.append({"marker": "ping", 'result': {}})
+        obj.settings["monitoring-buffer-limit"] = 100
+        obj.prepare()
+        for i in range(1000):
+            mon = [{"ts": i, "source": "local", "cpu": float(i) / 1000 * 100, "mem": 2, "bytes-recv": 100, "other": 0}]
+            obj.monitoring_data(mon)
+            for source, buffer in iteritems(obj.monitoring_buffer.data):
+                self.assertLessEqual(len(buffer), 100)
 
 
 class TestBlazeMeterClientUnicode(BZTestCase):
@@ -217,3 +232,54 @@ class TestResultsFromBZA(BZTestCase):
         total = cumulative_['']
         percentiles_ = total[KPISet.PERCENTILES]
         self.assertEquals(1050, percentiles_['99.0'])
+
+
+class TestMonitoringBuffer(BZTestCase):
+    def to_rad(self, deg):
+        return deg * math.pi / 180
+
+    def test_harmonic(self):
+        ITERATIONS = 50
+        SIZE_LIMIT = 10
+        mon_buffer = MonitoringBuffer(SIZE_LIMIT)
+        for i in range(ITERATIONS):
+            cpu = math.sin(self.to_rad(float(i) / ITERATIONS * 180))
+            mon = [{"ts": i, "source": "local", "cpu": cpu}]
+            mon_buffer.record_data(mon)
+            self.assertLessEqual(len(mon_buffer.data['local']), SIZE_LIMIT)
+
+    def test_downsample_theorem(self):
+        # Theorem: average interval size in monitoring buffer will always
+        # be less or equal than ITERATIONS / BUFFER_LIMIT
+        mon_buffer = MonitoringBuffer(100)
+        for i in range(5000):
+            mon = [{"ts": i, "source": "local", "cpu": 1, "mem": 2, "bytes-recv": 100, "other": 0}]
+            mon_buffer.record_data(mon)
+            for source, buffer in iteritems(mon_buffer.data):
+                self.assertLessEqual(len(buffer), 100)
+                sizes = [item['interval'] for item in viewvalues(buffer)]
+                avg_size = float(sum(sizes)) / len(sizes)
+                expected_size = 5000 / 100
+                self.assertLessEqual(avg_size, expected_size * 1.20)
+
+    def test_sources(self):
+        mon_buffer = MonitoringBuffer(10)
+        for i in range(100):
+            mon = [
+                {"ts": i, "source": "local", "cpu": 1, "mem": 2, "bytes-recv": 100},
+                {"ts": i, "source": "server-agent", "cpu": 10, "mem": 20},
+            ]
+            mon_buffer.record_data(mon)
+            for source, buffer in iteritems(mon_buffer.data):
+                self.assertLessEqual(len(buffer), 10)
+
+    def test_unpack(self):
+        ITERATIONS = 200
+        SIZE_LIMIT = 10
+        mon_buffer = MonitoringBuffer(SIZE_LIMIT)
+        for i in range(ITERATIONS):
+            mon = [{"ts": i, "source": "local", "cpu": 1}]
+            mon_buffer.record_data(mon)
+        unpacked = sum(item['interval'] for item in viewvalues(mon_buffer.data['local']))
+        self.assertEqual(unpacked, ITERATIONS)
+