+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - add setting up of CWD on server side
  - fix problem with sending included configs into cloud
  - fix cumulative fail criteria processing order
+ - limit the amount of monitoring data BlazeMeter uploader accumulates
 
 ## 1.6.2 <sup>8 jun 2016</sup>
  - fix passfail-related regression crash